+++ crates/dxm-artifacts/src/lib.rs
@@ -12,13 +12,13 @@ pub mod github;
 pub mod jg;
 
 /// Downloads and installs the given installation version to the given directory
-/// path.
+/// path. Returns the installation download URL.
 pub fn install<S, P>(
     client: &Client,
     platform: &ArtifactsPlatform,
     version: S,
     path: P,
-) -> Result<(), Box<dyn Error>>
+) -> Result<String, Box<dyn Error>>
 where
     S: AsRef<str>,
     P: AsRef<Path>,
@@ -28,22 +28,22 @@ where
     fs_err::create_dir_all(path)?;
 
     let mut file = NamedTempFile::with_suffix(platform.archive_name())?;
-    download(client, platform, version, file.as_file_mut())?;
+    let url = download(client, platform, version, file.as_file_mut())?;
 
     log::trace!("extracting fxserver archive");
     platform.decompress(file.reopen()?, path)?;
 
-    Ok(())
+    Ok(url)
 }
 
 /// Downloads the given installation version archive, and writes it to the given
-/// writer.
+/// writer. Returns the installation download URL.
 pub fn download<S, W>(
     client: &Client,
     platform: &ArtifactsPlatform,
     version: S,
     mut writer: W,
-) -> Result<(), Box<dyn Error>>
+) -> Result<String, Box<dyn Error>>
 where
     S: AsRef<str>,
     W: Write,
@@ -54,8 +54,8 @@ where
     let url = platform.runtime_url(version, commit_sha);
 
     log::trace!("download fxserver archive");
-    let bytes = client.get(url).send()?.bytes()?;
+    let bytes = client.get(&url).send()?.bytes()?;
     writer.write_all(&bytes)?;
 
-    Ok(())
+    Ok(url)
 }