+++ src/commands/artifacts/update.rs
@@ -23,9 +23,9 @@ pub fn execute(args: &ArgMatches) -> Result<(), Box<dyn Error>> {
         .get_one::<PathBuf>("manifest-path")
         .expect("no manifest path");
 
-    let (manifest_path, mut manifest) = crate::util::manifest::find(manifest_path)?;
+    let (manifest_path, manifest) = crate::util::manifest::find(manifest_path)?;
 
-    crate::util::artifacts::update(manifest_path, &mut manifest)?;
+    crate::util::artifacts::update(manifest_path, &manifest)?;
 
     Ok(())
 }