+++ crates/dxm-manifest/CHANGELOG.md
@@ -9,14 +9,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [dxm unreleased]
 
-### Fixed
+### Added
 
-- TOML writing erroring for new documents.
+- Lockfile structures.
 
 ### Changed
 
+- **Breaking:** updated the default value for `Artifact::version` from an empty
+  string to `latest-jg`.
+- **Breaking:** updated `Artifact::channel` to parse the `version` field from
+  the source TOML, and return the value only if the version is a valid update
+  channel.
 - Updated dependencies.
 - TOML categories are now written individiually instead of overwriting the entire file.
+
+### Removed
+
+- **Breaking:** `Artifact::set_channel` and the `channel` field in
+  `Artifact::fill_toml_item` - use `Artifact::set_version` and the `version`
+  field instead, with a stringified channel.
+
+### Fixed
+
+- TOML writing erroring for new documents.
 
 ## 0.1.1 - 2025-07-20